Capacity and utilization (#4696)
This PR is a follow up to #4605 which adds views into capacity and
utilization both at the silo and system level.

API: 
|op|method|url|
|--|--|--|
|silo_utilization_list | GET    |  /v1/system/utilization/silos |
|silo_utilization_view | GET | /v1/system/utilization/silos/{silo} |
|utilization_view          |               GET  |    /v1/utilization |

I'm not entirely satisfied w/ the silo utilization endpoints. They could
be this instead:

|op|method|url|
|--|--|--|
|silo_utilization_list | GET    |  /v1/system/silos-utilization |
|silo_utilization_view | GET | /v1/system/silos/{silo}/utilization |

Also take special note of the views

```rust
// For the eyes of end users
/// View of the current silo's resource utilization and capacity
#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
pub struct Utilization {
    /// Accounts for resources allocated to running instances or storage allocated via disks or snapshots
    /// Note that CPU and memory resources associated with a stopped instances are not counted here
    /// whereas associated disks will still be counted
    pub provisioned: VirtualResourceCounts,
    /// The total amount of resources that can be provisioned in this silo
    /// Actions that would exceed this limit will fail
    pub capacity: VirtualResourceCounts,
}

// For the eyes of an operator
/// View of a silo's resource utilization and capacity
#[derive(Clone, Debug, Deserialize, Serialize, JsonSchema)]
pub struct SiloUtilization {
    pub silo_id: Uuid,
    pub silo_name: Name,
    /// Accounts for resources allocated by in silos like CPU or memory for running instances and storage for disks and snapshots
    /// Note that CPU and memory resources associated with a stopped instances are not counted here
    pub provisioned: VirtualResourceCounts,
    /// Accounts for the total amount of resources reserved for silos via their quotas
    pub allocated: VirtualResourceCounts,
}
```

For users in the silo I use `provisioned` and `capacity` as the
language. Their `capacity` is represented by the quota set by an
operator. For the operator `provisioned` is the same but `allocated` is
used to denote the amount of resources allotted via quotas.

---

Note: I had planned to add a full system utilization endpoint to this PR
but that would increase the scope. Instead will ship that API as a part
of the next release. We can calculate some version of the full system
utilization on the client by listing all the silos and their
utilization.

---------

Co-authored-by: Sean Klein <sean@oxide.computer>