Configure bash environment for NGZ root.
This removes the special case for the switch zone and deploys a
consistent profile and bashrc to all non-global zones.

While understanding this is subjective, improvements here are:
 * A switch zone prompt includes the local switch number (0/1)
   if it can be determined;
 * PATH is configured to include additional directories useful
   within the zone;
 * The hostname part of the prompt is truncated in zones which
   have UUIDs as part of their name;
 * Coloured prompt, as per the GZ.

Pretty much everyone has their own preferred prompt format, so
consensus is unlikely here, but this is a step forward in having
consistency and a better PATH. In the limit, nobody will be logging
into these zones outside of a development environment anyway.