Add a background task for update plan execution
This PR is the first step in creating a background task that is
capable of taking a `Blueprint` and then reifying that blueprint into
deployed or updated software. This PR uses the initial version of a
Blueprint introduced in #4804. A basic executor that sends the related
`OmicronZonesConfig` to the appropriate sled-agents for newly added
sleds was created.

A test is included that shows how a hypothetical planner for an
`add-sled` workflow will deploy Omicron zones in a manner similar to
RSS, where first the internal DNS zone is deployed and then the internal
DNS and NTP zones are deployed. Deployment alwyas contains all zones
expected to be running on the sled-agent. Any zones running that are not
included are expected to be shut down.