Add a background task for update plan execution
Note: I expect this to lead to some discussion before getting merged.

This PR is the first step in creating a background task that is capable
of taking a `Blueprint` and then reifying that blueprint into deployed
or updated software. This PR includes the initial version of a blueprint
that is only capable of describing deployable omicron zones. A basic
executor that sends the related `OmicronZonesConfig` to the appropriate
sled-agents for newly added sleds was created.

A test is included that shows how a hypothetical planner for an
`add-sled` workflow will deploy Omicron zones in a manner similar to
RSS, where first the internal DNS zone is deployed and then the internal
DNS and NTP zones are deployed. Deployment alwyas contains all zones
expected to be running on the sled-agent. Any zones running that are not
included are expected to be shut down.

This background task is not yet wired up into `nexus/src/app/background/
init.rs`, because I believe this depends on a planner being added that
is capable of providing Blueprints to drive the task forward. The first
planner can be super simple and triggered from an operator HTTP request,
but we probably still want a basic planner. One reason is that the
planner itself will almost certainly want to wait for timesync before it
proceeds to installing the rest of the zones.