Optimize OID lookup for user-defined types (#4735)
# Background

Looking up OIDs for ENUMs consumes a non-trival amount of time during
initialization, since they are synchronously fetched one-by-one.
Although this is amortized for long-lived connections, it incurs a cost
on all tests using CRDB, and can be optimized. Additionally, these OID
caches are per-connection, which incurs redundant work when we have a
connection pool (and we do, with bb8). See #4733 for additional context.

Feedback from the Diesel maintainers:

> I think it would be reasonable to share this information at pool
level...
>
> [OIDs are] currently queried by one type at the time as we argued that
it won't matter for long running applications. There you do the overhead
once per connection and then you just hit the cache. I don't think we
did really consider short running applications (like tests) for that.

# This PR

1. Performs a one-time-lookup of user-defined ENUM OIDs on the first
connection, and shares it between subsequent connections. This
pre-populates the cache used by Diesel.
2. Modifies all `SqlType` derives to ensure that the `schema` is
specified as `public` (this ensures a cache hit)
3. Adds a test to compare the "list of user-defined enums to look up"
with the current DB schema, and ensures they're in-sync.

# Results

I focused on optimizing `integration_tests::basic::test_ping`, as this
is very close to a "no-op" version of the `#[nexus_test]`. I wrote a
script to repeatedly execute this test in a loop (excluding the first
result, as a "warm-up" run to populate CRDB).

- Before: Average time over 10 runs: 10.42 seconds
- After: Average time over 10 runs: 9.47 seconds

That's a small difference, but hopefully it'll provide a dent over "all
our tests", and avoid additional user-defined types from slowing down
all tests.

Part of https://github.com/oxidecomputer/omicron/issues/4733