[tufaceous-lib] make DeserializedManifest public and serializable, fix serialization (#4781)
This is prep work for #4690. I'd like to add support for making changes
to a DeserializedManifest, then serializing it.

However, it turned out that the bytesize crate does not serialize bytes
correctly (https://github.com/hyunsik/bytesize/issues/40). To address
this, just use a u64 and write our own deserializer, using the alternative
parse-size crate.

Also add a test to ensure that serialization of the fake manifest
roundtrips correctly.