[tests] time out tests in test_all after a duration (#4780)
We're seeing some tests in `test_all` hang forever (#4779). Set a
reasonable
upper bound on test duration.

This will also cause stdout and stderr for failing tests to be printed.
Doing
so on SIGTERM in general is tracked at
https://github.com/nextest-rs/nextest/issues/1208.

Also, bump up the required nextest version to 0.9.64 to make use of the
`binary_id` predicate.