Provide mechanism for RPWs to execute sagas
This commit adds a `SagaRequest` channel that can be shared with
reliable persistent workflow (RPW) background tasks so that a saga can
be executed as part of those task's activation.

This functionality will be used as part of downstairs replacement: when
a physical disk is marked as no longer in use, a series of RPWs and
sagas will ensure that any Crucible resources that were allocated to
that physical disk are moved, and make any necessary volume changes.