[tufaceous-lib] add fake artifacts with an mtime of zero (#4823)
In #4690 I'd like to be able to generate a tufaceous repo twice and
expect that each artifact has the same hash.

Writing tests for this is a bit annoying at the moment because at the
moment the overall repo has a different hash when generated again
(because TUF bakes timestamps into its format), but each artifact has
the same hash. #4690 has other work going on which makes writing this
test quite simple, so I'll include the test there.