[update-engine] reuse parent_key_and_child_index from existing steps (#4858)
During a dogfood mupdate on 2024-01-18, I saw some really weird output
with wicket's `rack-update attach` that looked like:

```
[sled 8 00:20:48]   Running .... 12a 5s 1/3) Writing host phase 2 to slot B
```

The "5s" is all wrong -- the letter there is supposed to indicate, for an
engine that has one or more nested engines, the index of that nested
engine. So
for example, if a step 12 has two nested engines, they would be marked
"12a" and "12b".

"5s" indicates that that's the 19th nested engine for that step, and we
definitely have nowhere near 19 nested engines for a step anywhere in
wicketd.

This turned out to be because we weren't reusing child indexes from earlier
steps in the sequence. Fix that, and also add:

* tests which catch this issue
* some dev-only code to wicket which made it easy to debug this locally