Chore: Update OPTE to v0.27.214
This PR introduces fixes to TCP state machine handling, which has been
causing new TCP flows to hang whenever a source-port is reused under certain
circumstances. Most of the relevant details are outlined on
oxidecomputer/opte#444, but the gist is that OPTE is far more permissive
about TCP packets which are allowed to be sent/received by guests.