[nexus] add support for ingesting TUF repos (#4690)
Implement basic support for uploading TUF repos via an endpoint. The PR
looks pretty big but most of it is fairly mechanical addition and
removal (and much of it has to be done in one go because of internal
dependencies.)

Also include a few more changes:

- Move more code to update-common.
- Move the `by_hash` and `by_id` maps to update-common's
`UpdatePlanBuilder`.
- Remove old update-related code and database migrations that will be
replaced by newer blueprint design. (This is the vast majority of
deleted code, and ideally would be a separate PR except it's a bit
inconvenient to have a PR stack with multiple schema migrations.)

This PR **does not include** actually storing TUF repos and replicating
them among sleds. That work has been deprioritized for now, to instead
focus on sled addition and removal.