Better error message when there's no default IP pool (#4880)
Closes #4864 

This is a bad error message to get when the problem is that there is no
default IP pool configured for your current silo:

```
not found: ip-pool with id "Default pool for current silo"
```

"Default pool for current silo" is not an id, so why would we call it
one? This is better:

```
not found: default IP pool for current silo
```

This PR is just making that possible.