Serialize blueprints in the database (#4899)
This replaces the in-memory blueprint storage added as a placeholder in
#4804 with cockroachdb-backed tables. Both the tables and related
queries are _heavily_ derived from the similar tables in the inventory
system (particularly serializing omicron zones and their related
properties). The tables are effectively identical as of this PR, but we
opted to keep the separate because we expect them to diverge some over
time (e.g., inventory might start collecting additional per-zone
properties that don't exist for blueprints, such as uptime).

The big exception to "basically the same as inventory" is the
`bp_target` table which tracks the current (and past) target blueprint.
Inserting into this table has some subtleties, and we use a CTE to check
and enforce the invariants. This is the first diesel/CTE I've written;
it's based on other similar CTEs in Nexus, but I'd still appreciate a
particularly careful look there.
 
Fixes #4793.