Add `flake.nix` for Nix users
In order to build Omicron, including running `cargo check` or
`rust-analyzer`, a number of non-Rust dependencies must be present. This
commit adds a [Nix flake] file which users of the Nix package manager
can [use to enter a dev shell][nix-develop] that contains those
dependencies, making it easy for Nix and NixOS users to develop Omicron
on NixOS.

The Nix flake configured here creates a build environment with `clang`
as the C compiler toolchain, and ensures that the following libraries
are available:

- `libclang` (of a version matching the `clang` C toolchain)
- `libpq`
- `libxml2`
- `libxmlsec`
- `openssl`
- `sqlite`

In addition, it provides `pkg-config`, `cmake`, and `rustup`, and
ensures that the various environment variables that the Rust bindings
for these libraries use to find the libraries are set correctly.

This makes it possible to run `cargo check`, `cargo build`, and
`rust-analyzer` on NixOS.

I've also added to the `.envrc` file so that `direnv` will automatically
use the Nix flake's dev environment, if a `nix` binary is present on the
user's system and that `nix` binary supports flakes.

[Nix flake]: https://nixos.wiki/wiki/Flakes
[nix-develop]: https://nixos.wiki/wiki/Development_environment_with_nix-shell#nix_develop