Perform inventory fetch queries in batches (#4932)
When querying the database for inventory data, we read a large number of
rows from a lot of different tables. Therefore, we want to ensure that
there is always an upper bound on the amount of data read per query.
Currently, an upper bound is ensured using _limits_: we add a limit to
the query, and if more rows would be returned, we fail the query. This
means that we can ensure an upper bound. However, the issue with this is
that the behavior when the upper bound is reached is to *fail* --- this
means that it's possible to select an upper bound that's too small, and
upon hitting that upper bound, we just fail...and there's no way to know
whether the selected limit is big enough until we hit it at runtime.
Instead of using hard limits, it's preferable to use _pagination_. By
paginating these queries, we still set an upper bound on the number of
rows returned by a single query, but rather than failing if that bound
is reached, we instead perform multiple queries.

This branch changes inventory fetch queries to use pagination with a
batch size limit, rather than a hard limit on query size, using the
pagination utilities added in #4632. I've replaced the
`inventory_collection_read_best_effort` and
`inventory_collection_read_all_or_nothing` methods, which limit the size
of the query, with a new `inventory_collection_read_batched` method,
which uses the limit parameter as a maximum size per batch. I've then
changed the existing `inventory_get_latest_collection` method to use
`inventory_collection_read_batched` rather than
`inventory_collection_read_all_or_nothing`.

Closes #4629