Rework {Sp,Rot,HostPhase1}Updater tests to be less flaky  (#4922)
There were three tests that were mostly copy/pasted, that all tried to
tame an inherent race between two tokio tasks: one is delivering an
update to the simulated SP, and the other is polling that simulated SP
for the progress of that update. The taming was by throttling the
simulated SP to only reply to a single message at a time, then checking
some internal state to infer what kind of message the SP had just
received, and using that inference to check some invariants. The
inference was typically right, but as evidenced by the name of this PR,
not perfect.

I took a fairly big hammer to the Rot and HostPhase1 updater tests: they
now just loop waiting to see the "Complete" progress message, and no
longer try to single-step through messages. Under normal conditions
(e.g., not heavily-loaded-CI systems), it's likely these tests will skip
from "no progress" to "complete", but that's ok.

I tried to rework the SpUpdater test to keep the same "single-step
through messages" to carefully check the progress reports by adding more
details to the simulated SP, so the test can know exactly when the
specific message it was trying to infer (i.e., `SpUpdater`'s requests
for the update status) arrives. This should fix the flake; if it
doesn't, we can replace this one with the same impl that the Rot and
HostPhase1 tests now have.

Fixes https://github.com/oxidecomputer/omicron/issues/4911