Add a background task for update plan execution (#4891)
This PR is the first step in creating a background task that is capable
of taking a `Blueprint` and then reifying that blueprint into deployed
or updated software. This PR uses the initial version of a Blueprint
introduced in #4804. A basic executor that sends the related
`OmicronZonesConfig` to the appropriate sled-agents for newly added
sleds was created.

A background task that loads the target `Blueprint` from the database
and feeds it to the executor is also included, along with a test for each.