[wicketd] replace snafu with thiserror (#4950)
This is currently the only use of snafu in omicron (outside of
transitive
deps), and I figure it is simple enough that we can replace it with
thiserror,
reducing use of a direct dependency.

This was the first time I'd encountered snafu so I read a bit about it.
As far
as I understand, the main benefit of snafu is that it pushes you heavily
towards an error-per-module pattern. However, thiserror does permit that
pattern as well, and in practice it is only a little more verbose than
snafu to
do right (`map_err` vs `context`, though `snafu` introduces new types
that
aren't in the source code like `IoSnafu` and `ParseSnafu`).