nix flake: Use clickhouse from `nixpkgs` in flake
The Clickhouse binary downloaded by
`tools/install_builder_prerequisites.sh` doesn't work nicely on NixOS
due to dynamically loading a bunch of libraries in a way that `nix-ld`
doesn't seem to help with. Therefore, this branch updates the Nix dev
shell flake to depend on the pre-built patched Clickhouse package from
nixpkgs, instead.

The flake's `devShell` now has a `shellHook` which ensures the
Clickhouse binary and `config.xml` in the nix store is symlinked into
`out/clickhouse`, so all the test tools which depend on it being there
should still work. I'm not actually sure if this is necessary, or if
`clickhouse` just needs to be on the `$PATH`; if it just needs to be on
the `$PATH` then we can remove the `shellHook`.