Revert "WIP: placeholder to keep PR open"
This reverts commit a3d0f567b49245b9a6f0926baf4b4128dd911cd5.

Switching to "push" based approach instead of "pulling" from dendrite.
This will be a simpler implementation and can more easily be replaced with
a pull approach later than if we implemented the pull approach and
determined that we'd prefer the push approach. Another big motivator,
possibly the most important motivation, is that omicron has access to
important information that we'd have to plumb into dendrite, like switch
location and in the near future *rack* location, so it makes more sense
for this to be in omicron today.