blueprint planner: skip non-provisionable sleds
I think I missed this on #4959, but instead of tacking onto that I'm
opening this as a separate PR to confirm the guard is in the right
place.

If a sled has been marked as non-provisionable by an operator, I'm
pretty confident we should NOT choose that sled when we're picking sleds
for additional service zones (e.g., Nexus). I'm less confident on what
we should do about NTP and Crucible zones. The change in this PR will
still attempt to add an NTP zone, but will not attempt to add any
Crucible zones. Is that the right place for it? Or should we still
add Crucible zones to a non-provisionable sled (if it has zpools that
don't yet have Crucible zones)?