[sled-agent] Ignore zone order when checking for changes
This addresses one of two problems mentioned in #4990. sled-agent's `PUT
/omicron-zones` endpoint is expected to be idempotent, and it checks
that if it receives a zone config with the same generation as it
currently has, that the contents are identical. This check was just a
straight equality check between two `Vec`s, because we have been careful
on the Nexus side to always sort the zone list consistently, but on

When sled-agent receives a config with a new generation number,
sled-agent itself may reorder the zone list between receiving it from
Nexus and storing it in its ledger. This will cause the first `PUT` of
the new generation to succeed, but any subsequent `PUT`s with identical
contents to fail the equality check.

Instead of trying ensure that sled-agent sorts the zone list in its
ledger consistently with how Nexus sorts them, this PR changes the
comparison to explicitly sort both the incoming list and the list loaded
from the ledger prior to comparing them.