Don't constrain baseboard when adding sled (#4987)
We were artificially limiting how we added sleds to a rack by forcing
them to be of type `Baseboard::Gimlet`. Instead of constructing a
`Baseboard` inside nexus, we instead send down the serial and part
numbers and use those to uniquely identify the sled. We ignore whether
it's a PC or Gimlet as long as the ids match. This is similar to how the
inventory works and allows adding a sled to a rack on the falcon a4x2
testbed.