tools: use SHA256 in `ci_download_cockroachdb`
Depends on #4961.

Currently, the `tools/ci_download_cockroachdb` script uses MD5 as the
checksum for the Cockroachdb tarball. This is unfortunate for two
reasons:

1. Upstream Cockroachdb (and the corresponding Illumos build) publish
   SHA256 digests for these tarballs, rather than MD5s (see
   https://www.cockroachlabs.com/docs/releases/ and
   https://illumos.org/downloads/). Using SHA256 rather than MD5 digests
   should make updating to a new version easier.

2. Nix requires SHA256 checksums for files downloaded as build inputs.
   Currently, the Nix flake can use the SHA256 checksums for Maghemite
   and Dendrite from `tools/maghemite_mgd_checksums` and
   `tools/dendrite_stub_checksums`, meaning that updating these versions
   does not require manually changing the Nix flake. However, because we
   use MD5 rather than SHA256 checksums for Cockroachdb, updating the
   Cockroachdb version requires manually changing the version in the Nix
   flake, which is a shame, especially if someone unfamiliar with Nix
   has to do it...

This commit changes `tools/cockroachdb_checksums` and the corresponding
`tools/ci_download_cockroachdb` script to use SHA256 rather than MD5.
I've changed the Nix flake to read the hash from this file rather than
hard-coding it, so now, the cockroachdb version can be updated without
touching the flake.